Removed fixed sidebar option
Choosing the fixed sidebar option caused it to move to the left, regardless of the 'sidebar on left' setting. Fine if the latter was also chosen, but otherwise unacceptable.

In other words, the fixed sidebar option only worked if the sidebar was on the left.

Removed until I can figure out how to accommodate both features.